We made some simplications in the algorithms of threads. Thus, we eliminate the mechanic for distinct errors and the recreation of selenium drivers.
Another significative change is the timeout of threads, since now the main process 'pings'
the threads of save and resolve throught the buffers (sending zeros) we decided to apply a simple timeout
of empty queue to decide if the thread remains alive. It means that now the threads gonna end the execution
in the 1st occurence of a timeout exception (raised on empty lists).